backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.26.0
celery>=5.3.6
redis>=5.0.1
flower>=2.0.0
kombu>=5.3.0
google-generativeai>=0.3.2
google-genai
# 影片生成相關
Pillow>=10.0.0
edge-tts>=6.1.0
replicate>=0.25.0
# 排程引擎
pytz>=2024.1
APScheduler>=3.10.0
# 社群平台 OAuth
aiohttp>=3.9.0
# 雲端儲存 (R2/S3)
boto3>=1.34.0
# 郵件服務
sendgrid>=6.10.0
# 認證相關
pyotp>=2.9.0
# 系統監控
psutil>=5.9.0
# Google Analytics
google-analytics-data>=0.18.0
google-auth>=2.25.0